Fixes 3D bar chart example.
It was using a CategoryAxis instead of Category3DAxis.

This error only triggers from ExtJS when running in debug mode, and this
was triggered because now the project switches Ext.NET script debugging
on by default if the EE is run with debug mode enabled (9e55e50f5ccd04).

This addresses issue #15 item 22.