Reviews the MenuPanel example to illustrate a new functionality.
This adds a CSS class and uses it on the first menu accordion,
illustrating the new SelectedCls and SelectedTextCls functionality
introduced on Ext.NET 4.3.0 via extnet/Ext.NET#1518.

This also adds IDs for easier handling of the individual menus
throughout the example. This easier console access is useful for debugging and
testing.